pyproject.toml
[tool.poetry]
name = "RollingForests"
version = "0.1.0"
description = ""
authors = ["WillyKess <williamhtkessner@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10 <4.0.0"
stable-baselines3 = {extras = ["mpi"], version = "^1.5.0"}
numpy = "^1.22.4"
gym = "^0.21"
selenium = "^4.2.0"
Pillow = "^9.1.1"
progressbar2 = "^4.0.0"
PyVirtualDisplay = "^3.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
